requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
